C01.01.01 - build(src): Reorganiztion and new library for specific data types
Reorganization of compile order. COMMS and IO are now on to of the order
and CONFIG and CROSSBAR are secondary packages dedicated for each board.
In case of a new system size (<7,8) the COMMS library becomes a local board
version. A new data types library was added to simplify the use of new
modules.